# Get rid of conversions in UnitRange iterator