@haampie 2018-03-22
Turns out I found a very contrived counter example where the conversion is actually necessary.

```
for i = false : true
  println(i)
end
```

would result in

```
false
1
```

after this PR :(.