@mbauman 2018-03-22
I'm surprised that this change fixed your performance given that the only method you actually hit is `done`… and these conversions (in just these methods by themselves) should be no-ops.  The start and next methods involved in this case are [here](https://github.com/JuliaLang/julia/pull/26572/files#diff-c5053a63cd3df77d9079e95f29f350adL469), just below your diff — that's what's causing all the conversions.  It looks like it was intentional to avoid overflow (see https://github.com/JuliaLang/julia/pull/13865).