@haampie 2018-03-22
You're right, this PR should be closed cause it's not addressing the actual issue.